Chore: changing color scheme
