+++ lib/common_widgets/widgets/cards/expansion_faq_tile.dart
@@ -21,7 +21,7 @@
 //         expandedColor: Colors.white,
 //         baseColor: Colors.white,
 //         leading: const CircleAvatar(
-//           backgroundColor: Colors.green,
+//           backgroundColor: Colors.purple,
 //           child: Text("FAQ"),
 //         ),
 //         title: Text(title),
